+++ NEWS.md
@@ -24,7 +24,9 @@
 to allow vectorization for `phi1` function in R/cch.R
 and added unit test to "tests/testthat/test-special-functions.R"
 
-* fixed  [issue #30](https://github.com/merliseclyde/BAS/issues/30) added n as hyperparameter if NULL and coerced to be a REAL for `intrinsic` prior 
+* fixed  [issue #30](https://github.com/merliseclyde/BAS/issues/30) added n as hyperparameter if NULL and coerced to be a REAL for `intrinsic` prior in `bas.glm`; added unit-test
+
+* fixed  [issue #39](https://github.com/merliseclyde/BAS/issues/39) added n as hyperparameter if NULL and coerced to be a REAL for `beta.prime` prior in `bas.glm`; added unit-test
 
 * fixed  [issue #27](https://github.com/merliseclyde/BAS/issues/27) where expected shrinkage with the JZS prior was greater than 1.  Added unit test.